flink/requirements.txt
psycopg2-binary==2.9.9
pyodbc==5.1.0
kafka-python==2.0.2
apache-flink==1.17.1
python-dotenv==1.0.1
pandas==1.3.5
confluent_kafka[avro]==1.9.2
avro-python3==1.8.1